# drmaatic/static/swagger.yml
openapi: 3.1.0
info:
  title: |
    DRMAAtic
  version: 1.0.0
  description: |
    <img src="/static/img/logo.png" alt="drmaatic logo" height="70">&nbsp;&nbsp;&nbsp;&nbsp;
    <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/3/3a/Slurm_logo.svg/839px-Slurm_logo.svg.png" alt="slurm logo" height="70">
    
    ### Introduction to DRMAAtic API
    
    The DRMAAtic API offers a comprehensive set of endpoints to facilitate authentication, job management, and task
    retrieval within the DRMAAtic system. DRMAAtic is a powerful platform designed to streamline and manage various tasks
    that should run on a DRM. The current implementation of DRMAAtic supports the SLURM DRM, but it can be easily
    extended to support other DRMs, given that they support the DRMAA specifications.
    
    #### DRMAAtic Overview

    DRMAAtic serves as a versatile platform with a range of capabilities, including:
    
    1. **User Authentication**: The API provides endpoints for authenticating users with DRMAAtic using JWT (JSON Web Tokens).
        Users can obtain JWT tokens by providing access tokens from external social authentication providers through OAuth2.
    
    2. **Job Management**: DRMAAtic enables users to create, manage, and monitor jobs seamlessly. Users can define jobs
        with varying parameters and types, submit them to the DRM (Digital Rights Management), and access job-related information.
    
...
    
...
    
...
    
    #### Key API Endpoints
    
...
    
...
    
...
    
...
servers:
...
security:
  - DRMAAticJWTAuth: [ ]
paths:
...
  /job/:
    post:
      operationId: createJob
      summary: 'Create a new job'
      description: |
        Create a new job and send it to the DRM.
        
        **The parameters of the job depend form the task that the user wants to execute.**
        Parameters can be of different type (string, integer, float, file). 
        To find out which type of parameters are needed for a specific task look at `/task/{name}`
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Job'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Job'
      responses:
        '201':
          description: 'Job successfully created'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Job'
                  - $ref: '#/components/schemas/SuperuserRequestedJob'
              examples:
                Job:
                  $ref: '#/components/examples/Job'
                SuperJob:
                  $ref: '#/components/examples/SuperuserRequestedJob'
        '404':
          description: 'Task not found'
          $ref: '#/components/responses/TaskNotFound'
        '400':
          description: 'The job could not be created'
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
                    example: 'The job could not be created'
      tags:
        - 'Job'
    get:
      operationId: listJobs
      summary: 'List all the jobs'
      description: 'List all the jobs that are owned by the user. If the user is a superuser, all the jobs are returned. This will also update the satus of the jobs in the database with the current status in the DRM'
      parameters:
        - name: page
          required: false
          in: query
          description: A page number within the paginated result set.
          schema:
            type: integer
        - name: page_size
          required: false
          in: query
          description: Number of results to return per page.
          schema:
            type: integer
        - name: status
          required: false
          in: query
          description: 'Filter by job status'
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: job_name
          required: false
          in: query
          description: 'Filter by the name of the task that was executed by the job'
          schema:
            type: string
        - name: search
          required: false
          in: query
          description: A search term.
          schema:
            type: string
      responses:
        '200':
          description: 'List of jobs matching the query. Only jobs that are owned by the user are returned, unless the user is a superuser'
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=4
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://api.example.org/accounts/?page=2
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SuperuserRequestedJob'
        '204':
          description: 'No jobs found. This is returned when the user is not a superuser or the user is not authenticated'
      tags:
        - 'Job'
  /job/{uuid}/:
    get:
      summary: 'Retrieve a specific job'
      operationId: retrieveJob
      description: Retrieve the job and if the current status in the database is not final, update it with the current status in the DRM
      parameters:
        - name: uuid
          in: path
          required: true
          description: 'UUID of the job'
          schema:
            type: string
      responses:
        '200':
          description: 'Job successfully retrieved'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Job'
                  - $ref: '#/components/schemas/SuperuserRequestedJob'
              examples:
                Job:
                  $ref: '#/components/examples/Job'
                SuperJob:
                  $ref: '#/components/examples/SuperuserRequestedJob'
        '404':
          description: 'Job not found'
          $ref: '#/components/responses/JobNotFound'
      tags:
        - 'Job'
    delete:
      summary: 'Delete a specific job'
      operationId: destroyJob
      description: Delete the job, stopping it in DRM, but preserving it in the database. Only superusers will be able to see the job after it has been deleted
      parameters:
        - name: uuid
          in: path
          required: true
          description: 'UUID of the job to be deleted'
          schema:
            type: string
      responses:
        '204':
          description: 'Job deleted'
        '403':
          description: 'You do not have permission to delete this job'
        '404':
          description: 'Job not found'
          $ref: '#/components/responses/JobNotFound'
      tags:
        - 'Job'
...
components:
  securitySchemes:
    DRMAAticJWTAuth:
      scheme: bearer
      bearerFormat: JWT
      in: header
      type: http
      description: 'The JWT authorization token that is given by DRMAAtic for the social provider access token'
    AccessTokenFromSocial:
      scheme: bearer
      bearerFormat: UUID
      in: header
      type: http
      description: 'The access token provided upon authentication with the social provider'
    OAuth2:
      type: oauth2
      description: 'OAuth2 authentication for the social provider. This will retrieve the access token from the social provider and this can be used to authenticate with DRMAAtic'
      flows:
        implicit:
          authorizationUrl: https://sandbox.orcid.org/oauth/authorize
          scopes:
            openid: Grants access to openid
  responses:
...
    TaskNotFound:
      description: 'Not found'
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Task 'test' not found"
    JobNotFound:
      description: 'Not found'
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Not found."
  schemas:
    JobStatus:
      type: string
      enum:
        - job has been rejected from the ws
        - job has been received from the ws
        - job has been created and sent to the DRM
        - process status cannot be determined
        - job is queued and active
        - job is queued and in system hold
        - job is queued and in user hold
        - job is queued and in user and system hold
        - job is running
        - job is system suspended
        - job is user suspended
        - job was stopped by the user
        - job finished normally
        - job finished, but failed
    Job:
      type: object
      properties:
        uuid:
          description: 'UUID of the job'
          type: string
          format: uuid
          readOnly: true
        task:
          description: 'Name of the task that the job has to execute'
          type: field
        job_description:
          description: 'Description of the job given by the user'
          type: string
        parent_job:
          description: >
            UUID of the job where this job should be attached to. 
            If this job is attached to another job, it will be executed in the same directory of the parent job.
            Note that this does not imply any dependency between the two jobs, for that use the `dependencies` field.
          type: string
        descendants:
          description: 'List of jobs UUID that depend on this job, directly or indirectly'
          type: string
          readOnly: true
        dependencies:
          description: 'List of jobs UUID that this job depends on'
          type: array
          items:
            type: string
          writeOnly: true
        depends_on:
          description: 'List of jobs UUID that this job depends on'
          type: array
          items:
            type: string
          readOnly: true
        dependency_type:
          description: |
            Type of dependency between this job and the one the job depends on.
            - afterany: the job will be executed after all the jobs it depends on have finished, no matter if it failed or not
            - afterok: the job will be executed after all the jobs it depends on have finished successfully
            - afternotok: the job will be executed after all the jobs it depends on have finished, but at least one of them failed
          enum:
            - afterany
            - afterok
            - afternotok
          type: string
          nullable: true
        status:
          description: 'Status of the job'
          enum:
            - job has been rejected from the ws
            - job has been received from the ws
            - job has been created and sent to the DRM
            - process status cannot be determined
            - job is queued and active
            - job is queued and in system hold
            - job is queued and in user hold
            - job is queued and in user and system hold
            - job is running
            - job is system suspended
            - job is user suspended
            - job was stopped by the user
            - job finished normally
            - job finished, but failed
          readOnly: true
        files_name:
          description:
            Dictionary of the files that the job got as input, with the name of the file created inn the output
            directory as key and the name of the file that was passed as input as value
          type: object
          readOnly: true
        creation_date:
          description: 'Date of creation of the job'
          type: string
          format: date-time
          readOnly: true
        params:
          description: 'List of parameters passed to the job, with their name and value.'
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                readOnly: true
              value:
                type: string
                maxLength: 5000
            required:
              - value
          readOnly: true
        param1:
          type: string
          writeOnly: true
          description: Parameter specific to Task 'test'
        param2:
          type: integer
          writeOnly: true
          description: Another parameter specific to Task 'test'
        param3:
          type: boolean
          writeOnly: true
          description: Parameter specific to Task 'test2'
        param4:
          type: file
          writeOnly: true
          description: Another parameter specific to Task 'test2'
      required:
        - task
    SuperuserRequestedJob:
      allOf:
        - $ref: '#/components/schemas/Job'
      type: object
      properties:
        update_date:
          description: 'Date of the last update of the job. This is updated every time the status of the job is updated'
          type: string
          format: date-time
          readOnly: true
        deleted:
          description: 'If the job has been deleted, this will be true'
          type: boolean
          readOnly: true
        sender_ip_addr:
          description: 'IP address of the user that created the job'
          type: string
          readOnly: true
        drm_job_id:
          description: 'ID of the job in the DRM'
          type: string
          readOnly: true
        user:
          description: 'Username of the user that created the job'
          type: string
          readOnly: true
...
  examples:
    Job:
      summary: 'Example of a job'
      value: {
        "uuid": "d8f3e7ea-91b5-4848-af91-fd4378bfe6f2",
        "task": "test",
        "depends_on": [ ],
        "status": "job finished normally",
        "files_name": {
          "file_0.fasta": "test.fasta",
          "file_1.fasta": "test.fasta"
        },
        "creation_date": "2023-10-02T14:39:09.785085+02:00",
        "params": [
          {
            "name": "file",
            "value": "file_0.fasta"
          },
          {
            "name": "file",
            "value": "file_1.fasta"
          }
        ]
      }
    SuperuserRequestedJob:
      summary: 'Example of a job requested by a superuser'
      value: {
        "uuid": "d8f3e7ea-91b5-4848-af91-fd4378bfe6f2",
        "task": "test",
        "depends_on": [ ],
        "sender_ip_addr": "127.0.0.1",
        "status": "job finished normally",
        "deleted": false,
        "drm_job_id": "586",
        "files_name": {
          "file_0.fasta": "test.fasta",
          "file_1.fasta": "test.fasta"
        },
        "creation_date": "2023-10-02T14:39:09.785085+02:00",
        "update_date": "2023-10-02T14:39:09.801388+02:00",
        "params": [
          {
            "name": "file",
            "value": "file_0.fasta"
          },
          {
            "name": "file",
            "value": "file_1.fasta"
          }
        ]
      }
...